# docs/swagger.yaml
basePath: /
definitions:
  dto.AddItemRequest:
    properties:
      product_id:
        description: UserID    uint    `json:"user_id,omitempty"`
        type: string
      quantity:
        type: integer
      variant_id:
        type: string
    required:
    - product_id
    - quantity
    type: object
...
  dto.CreateReviewDTO:
    properties:
      comment:
        type: string
      product_id:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - product_id
    - rating
    type: object
...
  dto.ReviewResponseDTO:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  dto.UpdateReviewDTO:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    type: object
...
  dto.WishlistItemResponseDTO:
    properties:
      added_at:
        type: string
      discount:
        minimum: 0
        type: number
      discount_type:
        enum:
        - fixed
        - percentage
        - ""
        type: string
      merchant_id:
        type: string
      name:
        type: string
      product_id:
        type: string
      sku:
        type: string
      total_price:
        type: number
    type: object
  dto.WishlistResponseDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.WishlistItemResponseDTO'
        type: array
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for customer and merchant authentication in a multivendor platform
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Multivendor API
  version: "1.0"
paths:
...
  /review:
    post:
      consumes:
      - application/json
      description: Create a new review by user
      parameters:
      - description: Review input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReviewResponseDTO'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a review
      tags:
      - reviews
  /reviews/{id}:
    delete:
      description: Delete a review by ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete a review
      tags:
      - reviews
    get:
      description: Retrieve a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponseDTO'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a review by ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update an existing review by ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated review input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponseDTO'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update a review
      tags:
      - reviews
  /wishlist:
    get:
      consumes:
      - application/json
      description: Retrieve the user's wishlist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WishlistResponseDTO'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get the user's wishlist
      tags:
      - wishlist
    post:
      consumes:
      - application/json
      description: Add a product to the user's wishlist
      parameters:
      - description: Wishlist body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: product added to wishlist
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Add a product to the wishlist
      tags:
      - wishlist
  /wishlist/{productID}:
    delete:
      consumes:
      - application/json
      description: Remove a product from the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Remove a product from the wishlist
      tags:
      - wishlist
  /wishlist/{productID}/check:
    get:
      consumes:
      - application/json
      description: Check if a specific product is in the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Check if a product is in the wishlist
      tags:
      - wishlist
  /wishlist/clear:
    delete:
      consumes:
      - application/json
      description: Remove all products from the user's wishlist
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Clear the user's wishlist
      tags:
      - wishlist
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"